feat: refactor renderPlugin to be JupyterFrontEndPlugin
