+++ jupyterlab_polus_render/jupyterlab_polus_render/_frontend.py
@@ -6,4 +6,4 @@ Information about the frontend package of the widgets.
 """
 
 module_name = "jupyterlab_polus_render"
-module_version = "^1.0.0"
+module_version = "^1.1.0"
