+++ jupyterlab_polus_render/pyproject.toml
@@ -37,7 +37,7 @@ classifiers = [
 dependencies = [
     "ipywidgets>=8.0.0",
 ]
-version = "1.0.0"
+version = "1.1.0"
 
 [project.optional-dependencies]
 docs = [
@@ -112,7 +112,7 @@ file = [
 ]
 
 [tool.tbump.version]
-current = "1.0.0"
+current = "1.1.0"
 regex = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)((?P<channel>a|b|rc|.dev)(?P<release>\\d+))?"
 
 [tool.tbump.git]
