+++ jupyterlab_polus_render/src/plugin.ts
@@ -1,21 +1,31 @@
-import { Application, IPlugin } from '@lumino/application';
-
+import { DOMWidgetView } from '@jupyter-widgets/base';
+import { Application } from '@lumino/application';
 import { Widget } from '@lumino/widgets';
-
 import { IJupyterWidgetRegistry } from '@jupyter-widgets/base';
-
-import * as widgetExports from './widget';
-
+import { JupyterFrontEndPlugin } from '@jupyterlab/application';
+import { PageConfig } from '@jupyterlab/coreutils';
+import { IFileBrowserFactory } from '@jupyterlab/filebrowser';
+import { store } from '@labshare/polus-render';
+import { RenderModel } from './widget';
 import { MODULE_NAME, MODULE_VERSION } from './version';
 
+
 const EXTENSION_ID = 'jupyterlab_polus_render:plugin';
+
+// Get the base URL of the JupyterLab session
+const baseUrl = PageConfig.getBaseUrl();
+// URL for serving images
+const renderFilePrefix = 'jupyterlab-polus-render/image'
 
 /**
  * The render plugin.
  */
-const renderPlugin: IPlugin<Application<Widget>, void> = {
+const renderPlugin: JupyterFrontEndPlugin<void> = {
   id: EXTENSION_ID,
-  requires: [IJupyterWidgetRegistry],
+  requires: [
+    IJupyterWidgetRegistry,
+    IFileBrowserFactory
+  ],
   activate: activateWidgetExtension,
   autoStart: true,
 }
@@ -27,11 +37,54 @@ export default renderPlugin;
  */
 function activateWidgetExtension(
   app: Application<Widget>,
-  registry: IJupyterWidgetRegistry
+  registry: IJupyterWidgetRegistry,
+  browserFactory: IFileBrowserFactory
 ): void {
+  const RenderView = class extends DOMWidgetView {
+    render() {
+      let imagePath = this.model.get('imagePath');
+      let overlayPath = this.model.get('overlayPath');
+      let isImagePathUrl = this.model.get('is_imagePath_url');
+      let isOverlayPathUrl = this.model.get('is_overlayPath_url');
+      let imageUrl = isImagePathUrl ? imagePath : `${baseUrl}${renderFilePrefix}${imagePath}`; // T/F condition ? valueIfTrue : valueIfFalse
+      let overlayUrl = isOverlayPathUrl ? overlayPath : `${baseUrl}${renderFilePrefix}${overlayPath}`;
+
+      // Set the image url
+      store.setState({
+        urls: [
+          imageUrl,
+        ],
+      });
+
+      // Set the overlay url
+      fetch(overlayUrl).then((response) => {
+        response.json().then((overlayData) => {
+          store.setState({
+            overlayData,
+          });
+          const heatmapIds = Object.keys(overlayData.value_range)
+            .map((d: any) => ({ label: d, value: d }))
+            .concat({ label: 'None', value: null });
+
+          store.setState({
+            heatmapIds,
+          });
+        });
+      });
+
+      this.el.innerHTML = `
+      <div style="width:100%;height:900px">
+      <polus-render></polus-render>
+      `;
+    }
+  }
+
   registry.registerWidget({
     name: MODULE_NAME,
     version: MODULE_VERSION,
-    exports: widgetExports,
+    exports: {
+      RenderModel,
+      RenderView
+    },
   });
 }
