+++ jupyterlab_polus_render/src/widget.ts
@@ -2,21 +2,15 @@
 
 import {
   DOMWidgetModel,
-  DOMWidgetView,
   ISerializers,
 } from '@jupyter-widgets/base';
 
 import { MODULE_NAME, MODULE_VERSION } from './version';
-import { PageConfig } from '@jupyterlab/coreutils';
-import { store } from '@labshare/polus-render';
 
 // Import the CSS
 import '../css/widget.css';
 
-// Get the base URL of the JupyterLab session
-const baseUrl = PageConfig.getBaseUrl();
-// URL for serving images
-const renderFilePrefix = 'jupyterlab-polus-render/image'
+
 
 export class RenderModel extends DOMWidgetModel {
   defaults() {
@@ -42,48 +36,4 @@ export class RenderModel extends DOMWidgetModel {
   static view_name = 'RenderView'; // Set to null if no view
   static view_module = MODULE_NAME; // Set to null if no view
   static view_module_version = MODULE_VERSION;
-}
-
-export class RenderView extends DOMWidgetView {
-  render() {
-    let imagePath = this.model.get('imagePath');
-    let overlayPath = this.model.get('overlayPath');
-    let isImagePathUrl = this.model.get('is_imagePath_url');
-    let isOverlayPathUrl = this.model.get('is_overlayPath_url');
-    let imageUrl = isImagePathUrl ? imagePath : `${baseUrl}${renderFilePrefix}${imagePath}`; // T/F condition ? valueIfTrue : valueIfFalse
-    let overlayUrl = isOverlayPathUrl ? overlayPath : `${baseUrl}${renderFilePrefix}${overlayPath}`;
-
-    // Set the image url
-    store.setState({
-      urls: [
-        imageUrl,
-      ],
-    });
-
-    // Set the overlay url
-    fetch(overlayUrl).then((response) => {
-      response.json().then((overlayData) => {
-        store.setState({
-          overlayData,
-        });
-        const heatmapIds = Object.keys(overlayData.value_range)
-          .map((d: any) => ({ label: d, value: d }))
-          .concat({ label: 'None', value: null });
-    
-        store.setState({
-          heatmapIds,
-        });
-      });
-    });    
-
-    this.el.innerHTML = `
-    <div style="width:100%;height:900px">
-    <div style="position: absolute; z-index: 100; right: 0">
-              <image-menu-web-component></image-menu-web-component>
-              <overlay-menu-web-component></overlay-menu-web-component>
-            </div>
-            <viv-viewer-web-component-wrapper></viv-viewer-web-component-wrapper>
-            </div>
-    `;
-  }
 }
